Use OpenAI compatible Gemini API
